backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langgraph==0.2.34
langchain-anthropic==0.2.1
wikipedia==1.4.0
python-dotenv==1.0.1
asyncio
aiohttp